prease heroku

+++ db/migrate/20171116012108_create_pokemon.rb
@@ -4,8 +4,8 @@ class CreatePokemon < ActiveRecord::Migration
       t.string :name
       t.string :type_1
       t.string :type_2
-      t.decimal :height
-      t.decimal :weight
+      t.integer :height
+      t.integer :weight
       t.integer :level
       t.string :nickname
       t.integer :trainer_id
